# Fix #1493 router loop for param routes