This pull request fixes issue #1493, which refers to a router loop recently introduced with commit 5bf6888 to fix param values. The router Find function will now correctly set the parameters and avoid the endless routing loop.

An additional test for the loop has been added (it is reproducable) and some router tests have been corrected to use a fresh context for the tests (to avoid leaking state). 

### Benchmark

Benchmarks are looking good, only minimal changes (~1% for static route performance)

With this patch:

```
rl@slimboy branch:bugfix/1493-# /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench '^(BenchmarkRouterStaticRoutes|BenchmarkRouterGitHubAPI|BenchmarkRouterParseAPI|BenchmarkRouterGooglePlusAPI)$'
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRouterStaticRoutes-8    	  91005	    11681 ns/op	      0 B/op	      0 allocs/op
BenchmarkRouterGitHubAPI-8       	  48444	    24682 ns/op	      1 B/op	      0 allocs/op
BenchmarkRouterParseAPI-8        	 293914	     3933 ns/op	      0 B/op	      0 allocs/op
BenchmarkRouterGooglePlusAPI-8   	 187198	     6247 ns/op	      0 B/op	      0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	5.083s
```

Before the patch (master@75620e6, some commits after v4.1.14):

```
rl@slimboy branch:master# /usr/bin/go test -benchmem -run=^$ github.com/labstack/echo/v4 -bench '^(BenchmarkRouterStaticRoutes|BenchmarkRouterGitHubAPI|BenchmarkRouterParseAPI|BenchmarkRouterGooglePlusAPI)$'
goos: linux
goarch: amd64
pkg: github.com/labstack/echo/v4
BenchmarkRouterStaticRoutes-8    	  93324	    11502 ns/op	      0 B/op	      0 allocs/op
BenchmarkRouterGitHubAPI-8       	  46220	    25976 ns/op	      1 B/op	      0 allocs/op
BenchmarkRouterParseAPI-8        	 276850	     3974 ns/op	      0 B/op	      0 allocs/op
BenchmarkRouterGooglePlusAPI-8   	 193024	     6177 ns/op	      0 B/op	      0 allocs/op
PASS
ok  	github.com/labstack/echo/v4	5.075s
```