@lammel 2020-02-18
@vishr Any chance this can be merged?
It has already bitten some people upgrading to the latest version.