@lammel 2020-02-18
Rebased the branch on current master.